Moved general field types to dedicated model sources
